add a calculation of percent complete to the participant model, move some helper classes out into their own files to avoid circular imports, add tests for participant name and percent complete
