+++ backend/app/model/participant_relationship.py
@@ -1,0 +1,16 @@
+import enum
+
+
+class Relationship(enum.Enum):
+    self_participant = 1
+    self_guardian = 2
+    dependent = 3
+    self_professional = 4
+
+    @classmethod
+    def has_name(cls, name):
+        return any(name == item.name for item in cls)
+
+    @classmethod
+    def options(cls):
+        return [item.name for item in cls]
