+++ backend/app/schema/schema.py
@@ -54,11 +54,12 @@ class ParticipantSchema(ModelSchema):
     class Meta:
         model = Participant
         fields = ('id', '_links', 'last_updated', 'name', 'relationship', 'user_id', 'avatar_icon', 'avatar_color',
-                  'has_consented', 'contact')
+                  'has_consented', 'contact', 'percent_complete')
     id = fields.Integer(required=False, allow_none=True)
     name = fields.Function(lambda obj: obj.get_name())
     relationship = EnumField(Relationship)
     user_id = fields.Integer(required=False, allow_none=True)
+    percent_complete = fields.Function(lambda obj: obj.get_percent_complete())
     _links = ma.Hyperlinks({
         'self': ma.URLFor('api.participantendpoint', id='<id>'),
         'user': ma.URLFor('api.userendpoint', id='<user_id>')
